+++ pkg/compact/compact_test.go
@@ -6,8 +6,13 @@ package compact
 import (
 	"testing"
 
+	"github.com/go-kit/kit/log"
+	"github.com/oklog/ulid"
 	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/prometheus/tsdb"
+	"github.com/thanos-io/thanos/pkg/objstore"
 
 	"github.com/thanos-io/thanos/pkg/block/metadata"
 	"github.com/thanos-io/thanos/pkg/errutil"
@@ -124,4 +129,54 @@ func TestGroupMaxMinTime(t *testing.T) {
 
 	testutil.Equals(t, int64(0), g.MinTime())
 	testutil.Equals(t, int64(30), g.MaxTime())
+}
+
+func TestConcurrentGroups(t *testing.T) {
+	labels1 := metadata.Thanos{Labels: map[string]string{"foo": "bar"}, Downsample: metadata.ThanosDownsample{Resolution: 0}}
+	labels2 := metadata.Thanos{Labels: map[string]string{"foo2": "bar2"}, Downsample: metadata.ThanosDownsample{Resolution: 0}}
+	metas := map[ulid.ULID]*metadata.Meta{
+		// Metas for first tenant.
+		ULID(1): {BlockMeta: tsdb.BlockMeta{ULID: ULID(1), MinTime: 0, MaxTime: 1000}, Thanos: labels1},
+		ULID(2): {BlockMeta: tsdb.BlockMeta{ULID: ULID(2), MinTime: 0, MaxTime: 1000}, Thanos: labels1},
+		ULID(3): {BlockMeta: tsdb.BlockMeta{ULID: ULID(3), MinTime: 0, MaxTime: 1000}, Thanos: labels1},
+		ULID(4): {BlockMeta: tsdb.BlockMeta{ULID: ULID(4), MinTime: 1000, MaxTime: 2000}, Thanos: labels1},
+		ULID(5): {BlockMeta: tsdb.BlockMeta{ULID: ULID(5), MinTime: 1000, MaxTime: 2000}, Thanos: labels1},
+		ULID(6): {BlockMeta: tsdb.BlockMeta{ULID: ULID(6), MinTime: 1000, MaxTime: 2000}, Thanos: labels1},
+		ULID(7): {BlockMeta: tsdb.BlockMeta{ULID: ULID(7), MinTime: 2000, MaxTime: 3000}, Thanos: labels1},
+		ULID(8): {BlockMeta: tsdb.BlockMeta{ULID: ULID(8), MinTime: 2000, MaxTime: 3000}, Thanos: labels1},
+		ULID(9): {BlockMeta: tsdb.BlockMeta{ULID: ULID(9), MinTime: 2000, MaxTime: 3000}, Thanos: labels1},
+		// Metas for second tenant.
+		ULID(10): {BlockMeta: tsdb.BlockMeta{ULID: ULID(10), MinTime: 0, MaxTime: 1000}, Thanos: labels2},
+		ULID(11): {BlockMeta: tsdb.BlockMeta{ULID: ULID(11), MinTime: 1000, MaxTime: 2000}, Thanos: labels2},
+		ULID(12): {BlockMeta: tsdb.BlockMeta{ULID: ULID(12), MinTime: 2000, MaxTime: 3000}, Thanos: labels2},
+		ULID(13): {BlockMeta: tsdb.BlockMeta{ULID: ULID(13), MinTime: 3000, MaxTime: 4000}, Thanos: labels2},
+		ULID(14): {BlockMeta: tsdb.BlockMeta{ULID: ULID(14), MinTime: 3000, MaxTime: 4000}, Thanos: labels2},
+		ULID(15): {BlockMeta: tsdb.BlockMeta{ULID: ULID(15), MinTime: 3000, MaxTime: 4000}, Thanos: labels2},
+		ULID(16): {BlockMeta: tsdb.BlockMeta{ULID: ULID(16), MinTime: 4000, MaxTime: 5000}, Thanos: labels2},
+	}
+	logger := log.NewNopLogger()
+	blocksMarkedForDeletion := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
+	garbageCollectedBlocks := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
+	planner := NewTSDBBasedPlanner(logger, []int64{1000, 3000})
+	bkt := objstore.NewInMemBucket()
+	grouper := NewDefaultGrouper(logger, bkt, false, true, true, prometheus.NewRegistry(), blocksMarkedForDeletion, garbageCollectedBlocks, metadata.NoneFunc, planner)
+	groups, err := grouper.Groups(metas)
+	testutil.Ok(t, err)
+	testutil.Equals(t, 5, len(groups))
+	testutil.Equals(t, ULIDs(13, 14, 15), groups[0].IDs())
+	testutil.Equals(t, ULIDs(10, 11, 12), groups[1].IDs())
+	testutil.Equals(t, ULIDs(1, 2, 3), groups[2].IDs())
+	testutil.Equals(t, ULIDs(4, 5, 6), groups[3].IDs())
+	testutil.Equals(t, ULIDs(7, 8, 9), groups[4].IDs())
+}
+
+func ULID(i int) ulid.ULID { return ulid.MustNew(uint64(i), nil) }
+
+func ULIDs(is ...int) []ulid.ULID {
+	ret := []ulid.ULID{}
+	for _, i := range is {
+		ret = append(ret, ulid.MustNew(uint64(i), nil))
+	}
+
+	return ret
 }